Optimization of Observe output.
- move PatchPointData vector to Observe class
- add pointer of Observe object to assemblation classes
- remove EqData type
