+++ src/fields/equation_output.hh
@@ -17,6 +17,7 @@
 #include "fields/field_values.hh"  // for FieldValue, FieldValue<>::Scalar
 #include "io/output_time_set.hh"   // for OutputTimeSet
 #include "io/output_mesh.hh"
+#include "io/observe.hh"
 #include "system/exceptions.hh"    // for ExcStream, operator<<, DECLARE_EXC...
 #include "tools/time_marks.hh"     // for TimeMark, TimeMark::Type
 
@@ -37,8 +38,6 @@ template<unsigned int dim> class AssemblyOutputNodeData;
 template<unsigned int dim> class AssemblyObserveOutput;
 template< template<IntDim...> class DimAssembly> class GenericAssembly;
 template< template<IntDim...> class DimAssembly> class GenericAssemblyObserve;
-struct PatchPointData;
-typedef std::vector<PatchPointData> PatchPointVec;
 
 
 /**
@@ -115,11 +114,6 @@ public:
     /// Selects the error control field out of output field set according to input record.
     typename OutputMeshBase::ErrorControlFieldFunc select_error_control_field();
     
-    /// Getter from patch_point_data.
-    inline PatchPointVec &patch_point_data() {
-	    return patch_point_data_;
-    }
-
 private:
     /**
      * Creates output selection from the field set.
@@ -187,9 +181,6 @@ private:
     GenericAssembly< AssemblyOutputNodeData > * output_node_data_assembly_;
     GenericAssembly< AssemblyOutputNodeData > * output_corner_data_assembly_;
     GenericAssemblyObserve< AssemblyObserveOutput > * observe_output_assembly_;
-
-    /// Holds observe data of eval points on patch
-    PatchPointVec patch_point_data_;
 
 };
