+++ src/io/observe.hh
@@ -33,6 +33,30 @@ template <typename T> class ElementDataCache;
 
 
 
+/// Holds data of one eval point on patch (index of element and local coordinations).
+struct PatchPointData {
+    /// Default constructor
+    PatchPointData() {}
+
+    /// Constructor with data mebers initialization
+    PatchPointData(unsigned int elm_idx, arma::vec loc_coords)
+    : elem_idx(elm_idx), local_coords(loc_coords), i_quad(0), i_quad_point(0) {}
+
+    /// Copy constructor
+    PatchPointData(const PatchPointData &other)
+    : elem_idx(other.elem_idx), local_coords(other.local_coords),
+      i_quad(other.i_quad), i_quad_point(other.i_quad_point) {}
+
+    unsigned int elem_idx;        ///< Index of element
+    arma::vec local_coords;       ///< Local coords of point
+    unsigned int i_reg;           ///< Index of region (use during patch creating)
+    unsigned int i_quad;          ///< Index of quadrature (use during patch creating), i_quad = dim-1
+    unsigned int i_quad_point;    ///< Index of point in quadrature (use during patch creating)
+};
+typedef std::vector<PatchPointData> PatchPointVec;
+
+
+
 /**
  * Helper class stores base data of ObservePoint and allows to evaluate
  * the nearest point to input_point_.
@@ -270,6 +294,11 @@ public:
     template <typename T>
     OutputDataPtr prepare_compute_data(std::string field_name, double field_time, unsigned int n_rows, unsigned int n_cols);
 
+    /// Getter of patch_point_data.
+    inline PatchPointVec &patch_point_data() {
+	    return patch_point_data_;
+    }
+
 
 
 protected:
@@ -316,6 +345,9 @@ protected:
 
 	/// Index of actual (last) time in \p observe_values_time_ vector
 	unsigned int observe_time_idx_;
+
+    /// Holds observe data of eval points on patch
+    PatchPointVec patch_point_data_;
 
 	friend class ObservePointAccessor;
 };
